// src/generated/server/worldmonitor/emergency/v1/service_server.ts
export interface Shelter {
  id: string;
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  island: string;
  type: string;
  status: string;
  capacity: number;
  petsAllowed: boolean;
  adaAccessible: boolean;
}

export interface ListSheltersRequest {
  island: string;
}

export interface ListSheltersResponse {
  shelters: Shelter[];
}